Refactor/move Maya Setdress API
